+++ src/controllers/filters/parser/FilterParser.cpp
@@ -1,12 +1,13 @@
 #include "FilterParser.hpp"
 
 #include "Application.hpp"
+#include "common/Channel.hpp"
 #include "controllers/filters/parser/Types.hpp"
 #include "providers/twitch/TwitchIrcServer.hpp"
 
 namespace filterparser {
 
-ContextMap buildContextMap(const MessagePtr &m)
+ContextMap buildContextMap(const MessagePtr &m, chatterino::Channel *channel)
 {
     auto watchingChannel =
         chatterino::getApp()->twitch.server->watchingChannel.get();
@@ -83,9 +84,8 @@ ContextMap buildContextMap(const MessagePtr &m)
     };
     {
         using namespace chatterino;
-        auto channel = getApp()->twitch2->getChannelOrEmpty(m->channelName);
-        auto *tc = dynamic_cast<TwitchChannel *>(channel.get());
-        if (!channel->isEmpty() && tc)
+        auto *tc = dynamic_cast<TwitchChannel *>(channel);
+        if (channel && !channel->isEmpty() && tc)
         {
             vars["channel.live"] = tc->isLive();
         }
@@ -102,12 +102,6 @@ FilterParser::FilterParser(const QString &text)
     , tokenizer_(Tokenizer(text))
     , builtExpression_(this->parseExpression(true))
 {
-}
-
-bool FilterParser::execute(const MessagePtr &message) const
-{
-    auto context = buildContextMap(message);
-    return this->execute(context);
 }
 
 bool FilterParser::execute(const ContextMap &context) const